Checkout ya funcional por fín + corrección cosas varias
Está terrible el checkout, no tengo más tiempo de ponerme a corregirlo, lo siento mucho... se que es horrible, a mi también me sangran los ojos de ver el css

+++ Backend/eCommerce/Controllers/AddressController.cs
@@ -1,5 +1,4 @@
-﻿using eCommerce.Models.Database.Entities;
-using eCommerce.Models.Dtos;
+﻿using eCommerce.Models.Dtos;
 using eCommerce.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,6 +8,7 @@ namespace eCommerce.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
+[Authorize]
 public class AddressController : Controller
 {
     private readonly AddressService _service;
@@ -18,70 +18,57 @@ public class AddressController : Controller
         _service = service;
     }
 
-
     [HttpGet("{id}")]
     public async Task<AddressDto> GetByIdAsync(long id)
     {
         return await _service.GetByIdAsync(id);
     }
 
+    [HttpGet("Get_User_Addresses/{id}")]
+    public async Task<IActionResult> GetAllAddresses(long id)
+    {
+
+        Claim userClaimId = User.FindFirst("id");
+        if (userClaimId == null) return Unauthorized(new {Message = "Debe iniciar sesión para llevar a cabo esta acción"});
+
+        try
+        {
+            if (id <= 0) return BadRequest("El ID del usuario es inválido.");
+
+            IEnumerable<AddressDto> addresses = await _service.GetAllByUserIdAsync(id);
+
+            if (addresses == null || !addresses.Any()) return NotFound("No se encontraron direcciones para este usuario.");
+
+            return Ok(addresses);
+
+        }
+        catch (Exception ex) {return StatusCode(500, $"Error al obtener las direcciones: {ex.Message}");}
+    }
+
     [HttpPost("Insert_Address")]
     public async Task<IActionResult> InsertAddress([FromBody] AddressDto addressDto)
     {
-        if (addressDto == null)
-        {
-            return BadRequest("Se requiere la dirección.");
-        }
+        Claim userClaimId = User.FindFirst("id");
+        if (userClaimId == null) return Unauthorized(new {Message = "Debe iniciar sesión para llevar a cabo esta acción"});
+
+        if (addressDto == null) return BadRequest(new {Message = "Se requiere la dirección."});
 
         bool success = await _service.CreateAddressAsync(addressDto);
-
-        if (success)
-        {
-            return Ok("Dirección creada correctamente.");
-        }
-
-        return StatusCode(500, "Hubo un error al crear la dirección.");
+        if (!success) return StatusCode(500, "Hubo un error al crear la dirección.");
+        
+        return Ok(new {Message = "Dirección creada correctamente."});
     }
 
-    [HttpGet("Get_All_Addresses")]
-    public async Task<IActionResult> GetAllAddresses([FromQuery] long userId)
-    {
-        try
-        {
-            // Valida que se haya enviado el userId
-            if (userId <= 0)
-            {
-                return BadRequest("El ID del usuario es inválido.");
-            }
-
-            // Llama al servicio para obtener las direcciones del usuario
-            var addresses = await _service.GetAllByUserIdAsync(userId);
-
-            if (addresses == null || !addresses.Any())
-            {
-                return NotFound("No se encontraron direcciones para este usuario.");
-            }
-
-            return Ok(addresses);
-        }
-        catch (Exception ex)
-        {
-            return StatusCode(500, $"Error al obtener las direcciones: {ex.Message}");
-        }
-    }
-    [HttpDelete("Delete_Address")]
+    [HttpDelete("Delete_Address/{id}")]
     public async Task<IActionResult> DeleteAddress(long id)
     {
+        Claim userClaimId = User.FindFirst("id");
+        if (userClaimId == null) return Unauthorized(new {Message = "Debe iniciar sesión para llevar a cabo esta acción"});
+
+
         bool isDeleted = await _service.DeleteAddressAsync(id);
+        if (!isDeleted) return NotFound(new {Message = "Dirección no encontrada" });
 
-        if (!isDeleted)
-        {
-            return NotFound(new { Message = "Address not found" });
-        }
-
-        return Ok(new { Message = "Address deleted successfully" });
+        return Ok(new {Message = "Dirección eliminada correctamente" });
     }
-
-
-
 }
